2024 - 10 - 03 12:20
Finished mostly all the of the basics for all the HTMLs and CSS. But there are some improvements and changes needed for the website. For Home I have a navBar and picture and name and a title that says home. For About me I will be adding a video and description but the site website links do connect with each other. for Contact will be adding the form and a button to submit. And lastly I will be uploading files to my project page to show all the projects I have done. I am only able to update my github recently as there was an issue with my github and it kept crashing on me even if i uninstalled and re-downloaded. I have fixed the issue and github works properly now.